refactor: rename use_case files and update imports to use 'use_case' suffix
