+++ backend/src/application/usecase/close_activity_use_case.py
