+++ backend/src/application/usecase/join_activity_use_case.py
