+++ backend/src/config/dependencies.py
@@ -5,11 +5,11 @@ from src.adapters.output.activity_repository_firestore_adapter import (
     ActivityRepositoryFirestoreAdapter,
 )
 from src.adapters.output.activity_repository_mongo_adapter import ActivityRepositoryAdapter
-from src.application.usecase.close_activity_service import CloseActivityService
-from src.application.usecase.create_activity_service import CreateActivityService
-from src.application.usecase.evaluation_activity_service import EvaluationActivityService
-from src.application.usecase.get_activity_service import GetActivityService
-from src.application.usecase.join_activity_service import JoinActivityService
+from src.application.usecase.close_activity_use_case import CloseActivityService
+from src.application.usecase.create_activity_use_case import CreateActivityService
+from src.application.usecase.evaluation_activity_use_case import EvaluationActivityService
+from src.application.usecase.get_activity_use_case import GetActivityService
+from src.application.usecase.join_activity_use_case import JoinActivityService
 from src.config.settings import initialize_settings
 
 logger = logging.getLogger(__name__)
